refactor: move password-field core and base CSS to styles folder
