+++ packages/password-field/src/styles/vaadin-password-field-button-core-styles.js
