+++ packages/password-field/src/vaadin-password-field-button.js
@@ -10,7 +10,7 @@ import { DirMixin } from '@vaadin/component-base/src/dir-mixin.js';
 import { PolylitMixin } from '@vaadin/component-base/src/polylit-mixin.js';
 import { CSSInjectionMixin } from '@vaadin/vaadin-themable-mixin/css-injection-mixin.js';
 import { ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
-import { passwordFieldButton } from './vaadin-password-field-button-core-styles.js';
+import { passwordFieldButton } from './styles/vaadin-password-field-button-core-styles.js';
 
 /**
  * An element used internally by `<vaadin-password-field>`. Not intended to be used separately.
